Runner mega-update.
Added perf mode to run linux perf on the specified inputs. This
works basically exactly like bench mode. Perf files are found in
the timings folder, with the name perf_<fast|slow>_<k>_<n>.txt.

Added option -a to always say yes to prompts. This means that
reference files will always be generated when missing, and always
reused when present.

Added option -i to compile the fast using icc instead of gcc. So
far only tested under Windows.

The results and logfiles directories are now generated by the runner
and not by lda. This is (again) because of some silly cross-platform
issues that appeared with a recent Windows update.